% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIChawkes.R
\name{DIChawkes}
\alias{DIChawkes}
\title{Deviance Information Criterion
The deviance information criterion (DIC) is a hierarchical
modeling generalization of the Akaike information criterion (AIC).
It is particularly useful in Bayesian model selection problems where the posterior distributions of the models have been obtained by Markov chain Monte Carlo (MCMC) simulation. DIC is an asymptotic approximation as the sample size becomes large, like AIC. It is only valid when the posterior distribution is approximately multivariate normal.}
\usage{
DIChawkes(xdata, matW, matM, matQ)
}
\arguments{
\item{xdata}{data from a hawkes process}

\item{matW}{W parameter matrix resulting from MCMC}

\item{matM}{M parameter matrix resulting from MCMC}

\item{matQ}{Q parameter matrix resulting from MCMC}
}
\value{
DIC result
}
\description{
Deviance Information Criterion
The deviance information criterion (DIC) is a hierarchical
modeling generalization of the Akaike information criterion (AIC).
It is particularly useful in Bayesian model selection problems where the posterior distributions of the models have been obtained by Markov chain Monte Carlo (MCMC) simulation. DIC is an asymptotic approximation as the sample size becomes large, like AIC. It is only valid when the posterior distribution is approximately multivariate normal.
}
