% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodhawkes.R
\name{likelihoodhawkes}
\alias{likelihoodhawkes}
\title{Estimação de Verossimilhança do Simulador do processo de Hawkes}
\usage{
likelihoodhawkes(dados2, M, W, Q, horizon)
}
\arguments{
\item{dados2}{dados do processo}

\item{M}{Vetor de processos Gaussianos relacionado a lambda}

\item{W}{Vetor de processos Gaussianos relacionado a alpha}

\item{Q}{Vetor de processos Gaussianos relacionado a beta}

\item{horizon}{número positivo, horizonte dos dados.}
}
\description{
Esta é a função para estimar a função de Verossimilhança dos dados a partir de um processo de Hawkes
}
