% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMChawkes.R
\name{MCMChawkes}
\alias{MCMChawkes}
\title{Algoritmo MCMC}
\usage{
MCMChawkes(
  point,
  xdata,
  xcov,
  iter = 1e+05,
  bar = 90000,
  pul = 2,
  AMpri = matrix(rep(1, nrow(point)), ncol = 1),
  BMpri = 100 * diag(1, nrow(point)),
  AQpri = matrix(rep(1, nrow(point)), ncol = 1),
  BQpri = 100 * diag(1, nrow(point)),
  AW = as.matrix(rep(0, ncol(xcov))),
  BW = 100 * diag(1, ncol(xcov)),
  avW = 1,
  bvW = 1,
  avM = 1,
  bvM = 1,
  avQ = 1,
  bvQ = 1,
  abW = 0.01,
  bbW = 0.01,
  abM = 0.01,
  bbM = 0.01,
  abQ = 0.01,
  bbQ = 0.01
)
}
\arguments{
\item{point}{Malha de pontos}

\item{xdata}{Matriz de dados simulados de um processo de Hawkes}

\item{xcov}{Matriz de covariáveis}

\item{iter}{número inteiro positivo, interações do algoritmo MCMC}

\item{bar}{número inteiro positivo, período de burn-in do algoritmo MCMC}

\item{pul}{número inteiro positivo, pulo nas interações do algoritmo MCMC}

\item{AMpri, AQpri}{vetor de prioris A para o parâmetros M e Q, respectivamente}

\item{BMpri, BQpri}{matriz de prioris B para o parâmetros M e Q, respectivamente}

\item{AW, BW}{matriz de prioris B para o parâmetros PSI lambda}

\item{avW, avM, avQ}{número inteiro positivo, priori do valor A do parâmetro v_lambda, v_alpha e v_beta}

\item{bvW, bvM, bvQ}{número inteiro positivo, priori do valor B do parâmetro v_lambda, v_alpha e v_beta}

\item{abW, abM, abQ}{número inteiro positivo, priori do valor A do parâmetro b_lambda, b_alpha e b_beta}

\item{bbW, bbM, bbQ}{número inteiro positivo, priori do valor B do parâmetro b_lambda, b_alpha e b_beta}
}
\description{
Esta é a função para realizar o algoritmo Monte Carlo via Cadeias de Markov para os parâmetros de um processo de Hawkes.
}
