% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prchawkes.R
\name{prchawkes}
\alias{prchawkes}
\title{Simulador do processo de Hawkes}
\usage{
prchawkes(
  n,
  PSI_lambda,
  b_lambda,
  v_lambda,
  mu_alpha,
  b_alpha,
  v_alpha,
  mu_beta,
  b_beta,
  v_beta,
  horizon
)
}
\arguments{
\item{n}{um número inteiro positivo, número de pontos na malha}

\item{PSI_lambda}{vetor de covariaveis px1}

\item{b_lambda}{número inteiro positivo}

\item{v_lambda}{número inteiro positivo, variância do parâmetro lambda}

\item{b_alpha}{número inteiro positivo}

\item{v_alpha}{número inteiro positivo, variância do parâmetro alpha}

\item{b_beta}{número inteiro positivo}

\item{v_beta}{número inteiro positivo, variância do parâmetro beta}

\item{horizon}{número positivo, horizonte dos dados.}

\item{PSI_alpha}{vetor de covariaveis px1}

\item{PSI_beta}{vetor de covariaveis px1}
}
\description{
Esta é a função para simuar os dados a partir de um processo de Hawkes
}
