% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodhawkes.R
\name{likelihoodhawkes}
\alias{likelihoodhawkes}
\title{Estimação de Verossimilhança do Simulador do processo de Hawkes}
\usage{
likelihoodhawkes(data0, lambda0, alpha0, beta0)
}
\arguments{
\item{data0}{dados do processo}

\item{lambda0}{Vetor de parâmetros do processo}

\item{alpha0}{Vetor de parâmetros do processo}

\item{beta0}{Vetor de parâmetros do processo}
}
\description{
Esta é a função para estimar a função de Verossimilhança dos dados a partir de um processo de Hawkes
}
